+++ eslint.config.js
@@ -1,24 +1,15 @@
-import js from '@eslint/js';
-import globals from 'globals';
+import { defineConfig } from "eslint/config";
 
-export default [
-    {
-        ...js.configs.recommended,
-        files: ['src/**/*.js']
+export default defineConfig({
+    files: ['src/**/*.js'],
+    rules: {
+        quotes: ['error', 'single'],
+        semi: ['error', 'always'],
+        curly: ['error', 'multi-line'],
+        'no-unused-vars': ['warn', { argsIgnorePattern: '^_' }]
     },
-    {
-        files: ['src/**/*.js'],
-        rules: {
-            quotes: ['error', 'single'],
-            semi: ['error', 'always'],
-            curly: ['error', 'multi-line'],
-        },
-        languageOptions: {
-            ecmaVersion: 2022,
-            sourceType: 'module',
-            globals: {
-                ...globals.node
-            }
-        }
+    languageOptions: {
+        ecmaVersion: 2022,
+        sourceType: 'module'
     }
-];
+});
