+++ package.json
@@ -31,16 +31,15 @@
     "src/"
   ],
   "devDependencies": {
-    "@babel/eslint-parser": "^7.28.5",
-    "@typescript-eslint/eslint-plugin": "^8.52.0",
-    "@typescript-eslint/parser": "^8.52.0",
-    "c8": "^10.1.3",
+    "@typescript-eslint/eslint-plugin": "^8.57.1",
+    "@typescript-eslint/parser": "^8.57.1",
+    "c8": "^11.0.0",
     "chai": "^6.2.2",
     "env-cmd": "^11.0.0",
-    "eslint": "^9.39.2",
+    "eslint": "^10.1.0",
     "mocha": "^11.7.5",
     "mocha-sonarqube-reporter": "^1.0.2",
-    "sinon": "^21.0.1"
+    "sinon": "^21.0.3"
   },
   "repository": {
     "type": "git",
